wrap-build: compare line numbers of errors to determine similarity
Check the line number, located beween the first two ":" symbols,
as an indication of whether the build succeeded or failed.

For every error, strip out everything but these line numbers.
Compare the resulting output, and if it's the same, then both
Arduino and Codebender failed in roughly the same way.

Signed-off-by: Sean Cross <sean@xobs.io>